services/chattts/requirements.txt
antlr4-python3-runtime==4.9.3
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
chattts-fork==0.0.8
click==8.1.8
colorama==0.4.6
einops==0.8.0
einx==0.3.0
encodec==0.1.1
filelock==3.13.1
flask==3.1.0
frozendict==2.4.6
fsspec==2024.2.0
huggingface-hub==0.27.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.3
markupsafe==2.1.5
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
packaging==24.2
pip==22.2.1
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.2
scipy==1.15.1
setuptools==63.2.0
sympy==1.13.1
tokenizers==0.19.1
torch==2.1.2
torchaudio==2.1.2
tqdm==4.67.1
transformers==4.41.2
typing-extensions==4.9.0
urllib3==2.3.0
uv==0.5.21
vector-quantize-pytorch==1.21.2
vocos==0.1.0
werkzeug==3.1.3
